backend/requirements/base.txt
Django==4.2.21
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
boto3==1.35.24
Pillow==10.4.0
python-dotenv==1.0.1
